requirements.txt
# PDF-to-DOCX Converter Requirements

# Core PDF processing
PyMuPDF>=1.23.0

# OCR capabilities
pytesseract>=0.3.10
tesseract>=0.1.3

# Word document generation
python-docx>=0.8.11

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Additional OCR engine (fallback)
easyocr>=1.7.0

# Utility libraries
numpy>=1.24.0
scipy>=1.11.0

# CLI and logging
click>=8.1.0
colorama>=0.4.6
tqdm>=4.65.0

# Configuration
pyyaml>=6.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
